FROMGIT: kbuild: Add __cc-option macro
cc-option uses KBUILD_CFLAGS and KBUILD_CPPFLAGS when it determines
whether an option is supported or not. This is fine for options used to
build the kernel itself, however some components like the x86 boot code
use a different set of flags.

Add the new macro __cc-option which is a more generic version of
cc-option with additional parameters. One parameter is the compiler
with which the check should be performed, the other the compiler options
to be used instead KBUILD_C*FLAGS.

Refactor cc-option and hostcc-option to use __cc-option and move
hostcc-option to scripts/Kbuild.include.

BUG=chromium:729743, chromium:702741
TEST=build for squawks

Conflicts:
	scripts/Kbuild.include

Change-Id: I8beb1fa33e746538a8196858b63ffc80935d47b3
Suggested-by: Arnd Bergmann <arnd@arndb.de>
Suggested-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Acked-by: Arnd Bergmann <arnd@arndb.de>
Acked-by: Michal Marek <mmarek@suse.com>
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
(cherry picked from git.kernel.org masahiroy/linux-kbuild kbuild
commit 9f3f1fd299768782465cb32cdf0dd4528d11f26b)
Reviewed-on: https://chromium-review.googlesource.com/548977
Reviewed-by: Guenter Roeck <groeck@chromium.org>